header_chain_prover: Move to it's own crate.
